# server/public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"SystemErrorCode": {
				"enum": [
					404,
					400,
					500,
					409
				],
				"type": "number"
			},
			"ISystemError": {
				"properties": {
					"error": {
						"$ref": "#/components/schemas/SystemErrorCode"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"error",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IProfile": {
				"properties": {
					"guid": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"secondName": {
						"type": "string"
					},
					"link": {
						"type": "string",
						"description": "Ссылка на мессенджер"
					},
					"phone": {
						"type": "string",
						"description": "Номер телефона для связи"
					},
					"symlinkedWith": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "Связан с организацией | GUID ORG"
					},
					"isActive": {
						"type": "boolean",
						"description": "Активен ли профиль"
					},
					"picture": {
						"type": "string"
					},
					"favouriteDevices": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"guid",
					"firstName",
					"secondName",
					"link",
					"phone",
					"symlinkedWith"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProfileCreateInterfaceRequestModel": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"secondName": {
						"type": "string"
					},
					"link": {
						"type": "string",
						"description": "Ссылка на мессенджер"
					},
					"phone": {
						"type": "string",
						"description": "Номер телефона для связи"
					},
					"guid": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"secondName",
					"link",
					"phone",
					"guid"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_ProfileCreateInterfaceRequestModel.Exclude_keyofProfileCreateInterfaceRequestModel.guid__": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"secondName": {
						"type": "string"
					},
					"link": {
						"type": "string",
						"description": "Ссылка на мессенджер"
					},
					"phone": {
						"type": "string",
						"description": "Номер телефона для связи"
					}
				},
				"required": [
					"firstName",
					"secondName",
					"link",
					"phone"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_ProfileCreateInterfaceRequestModel.guid_": {
				"$ref": "#/components/schemas/Pick_ProfileCreateInterfaceRequestModel.Exclude_keyofProfileCreateInterfaceRequestModel.guid__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"IRole": {
				"properties": {
					"guid": {
						"type": "string"
					},
					"deviceList": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"userList": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"adminList": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"ownerList": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"alias": {
						"type": "string"
					},
					"organizationGuid": {
						"type": "string"
					}
				},
				"required": [
					"guid",
					"deviceList",
					"userList",
					"adminList",
					"ownerList",
					"alias",
					"organizationGuid"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "header",
				"flow": "implicit",
				"scopes": {}
			}
		}
	},
	"info": {
		"title": "server",
		"version": "0.0.0",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/v1/profile/create": {
			"post": {
				"operationId": "CreateProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProfile"
								}
							}
						}
					}
				},
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProfileCreateInterfaceRequestModel"
							}
						}
					}
				}
			}
		},
		"/api/v1/profile/update": {
			"post": {
				"operationId": "UpdateProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProfile"
								}
							}
						}
					}
				},
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_ProfileCreateInterfaceRequestModel.guid_"
							}
						}
					}
				}
			}
		},
		"/api/v1/profile/identifyMyself": {
			"get": {
				"operationId": "GetProfileFromRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProfile"
								}
							}
						}
					}
				},
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/profile/identity": {
			"get": {
				"operationId": "GetProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProfile"
								}
							}
						}
					}
				},
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "guid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/profile/avatar": {
			"post": {
				"operationId": "SetAvatar",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProfile"
								}
							}
						}
					}
				},
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"b64": {
										"type": "string"
									}
								},
								"required": [
									"b64"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/profile/favourite": {
			"post": {
				"operationId": "AddFavourite",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProfile"
								}
							}
						}
					}
				},
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"guid": {
										"type": "string"
									}
								},
								"required": [
									"guid"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "RemoveFavourite",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProfile"
								}
							}
						}
					}
				},
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"guid": {
										"type": "string"
									}
								},
								"required": [
									"guid"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/profile/employee": {
			"get": {
				"operationId": "GetAllEmployees",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ISystemError"
										},
										{
											"items": {
												"$ref": "#/components/schemas/IRole"
											},
											"type": "array"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "orgGuid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}